+++ NEWS.md
@@ -2,6 +2,14 @@ Phan NEWS
 
 03 Apr 2018, Phan 0.12.5
 ------------------------
+New Features(CLI, Configs)
++ Warn about string and numeric literals that are no-ops. (E.g. `<?php 'notEchoedStr'; "notEchoed $x"; ?>`)
+  New issue types: `PhanNoopStringLiteral`, `PhanNoopEncapsulatedStringLiteral`, `PhanNoopNumericLiteral`.
+
+  Note: This will not warn about Phan's [inline type checks via string literals](https://github.com/phan/phan/wiki/Annotating-Your-Source-Code#inline-type-checks-via-string-literals)
++ Warn about invalid expressions/variables encapsulated within double-quoted strings or within heredoc strings.
+  New issue type: `TypeSuspicioousStringExpression` (May also emit `TypeConversionFromArray`)
+
 Plugins
 + Add an option `'php_native_syntax_check_max_processes'` to `'plugin_config'` for `InvokePHPNativeSyntaxCheckPlugin`.